requirements.txt
boto3==1.4.8
Django==1.11.7
django-storages==1.6.5
djangorestframework==3.7.3
zappa==0.45.1
